Implemented new sound engine (WIP)

+++ rust/pyxel-engine/src/audio.rs
@@ -117,7 +117,7 @@ impl Pyxel {
         &self,
         channel_index: u32,
         sequence: &[u32],
-        start_tick: Option<f64>,
+        start_tick: Option<u32>,
         should_loop: bool,
         should_resume: bool,
     ) {
@@ -142,7 +142,7 @@ impl Pyxel {
         &self,
         channel_index: u32,
         sound_index: u32,
-        start_tick: Option<f64>,
+        start_tick: Option<u32>,
         should_loop: bool,
         should_resume: bool,
     ) {
@@ -154,7 +154,7 @@ impl Pyxel {
         );
     }
 
-    pub fn playm(&self, music_index: u32, start_tick: Option<f64>, should_loop: bool) {
+    pub fn playm(&self, music_index: u32, start_tick: Option<u32>, should_loop: bool) {
         let num_channels = self.channels.lock().len();
         let musics = self.musics.lock();
         let music = musics[music_index as usize].lock();
